Implement rate limiting in toStimReq.
toStimReq should only output FiniteDurations if there has been a
sufficient change in the channels of the agent -- otherwise it should
discard the input. Note that the threshold check is done before
pruning away values outside of the sight range, such that an output
may consist of only None if the values that trigger an output all lie
outside this range.