Fix issue with resources in mock server. Details inside
When starting a server you get a Stream[F,Stream[F,Socket[F]]]
which can be interpreted two ways. The wrong interpretation is
that the outer stream represents opening a server, and the inner
stream is the resulting stream of incoming sockets.
The correct interpretation is that the outer stream is the stream
of connecting sockets, whereas the inner stream is a one-element
stream handling the socket as a resource. Upon closing the inner
stream this resource is then cleaned up, which is exactly what
happens when it is ran through a queue.

The solution in fs2 1.0 is to label the inner one element stream
as a Resource[F,Socket] which will make this more apparent.